fix login, property loading, delete and float type mismatches

+++ MobiFon.Services/Services/PropertyRatingService/PropertyRatingService.cs
@@ -64,7 +64,7 @@ namespace MobiFon.Services.Services.PropertyRatingService
 
         public void Update(PropertyRatingDto entity)
         {
-            float oldRating = entity.Rating;
+            double oldRating = entity.Rating;
             unitOfWork.PropertyRatingRepository.Update(entity);
 
             unitOfWork.SaveChanges();
@@ -80,7 +80,7 @@ namespace MobiFon.Services.Services.PropertyRatingService
 
         public async Task<PropertyRatingDto> UpdateAsync(PropertyRatingDto entity)
         {
-            float oldRating = entity.Rating;
+            double oldRating = entity.Rating;
             unitOfWork.PropertyRatingRepository.Update(entity);
 
             await unitOfWork.SaveChangesAsync();
